add service files and update database
