+++ frontend/src/App.jsx
@@ -6,7 +6,6 @@ import About from "./components/About";
 import Contact from "./components/Contact";
 import Services from "./components/Services";
 
-
 const App = () => {
   return (
     <Router>
