+++ frontend/src/services/api_services.jsx
@@ -1,15 +1,13 @@
-// Fetch hero and section data
-export const fetchHeroSectionData = async () => {
-  const response = await axios.get(
-    "https://localhost:62786/jsonapi/hero_section_data"
-  );
+const API_URL = "https://localhost:62786/jsonapi";
+
+// Fetch service page data
+export const fetchServicesData = async () => {
+  const response = await axios.get(`${API_URL}/node/services`);
   return response.data;
 };
 
 // Fetch services cards
 export const fetchServiceCards = async () => {
-  const response = await axios.get(
-    "https://localhost:62786/jsonapi/service_cards"
-  );
+  const response = await axios.get(`${API_URL}/hero_section_data`);
   return response.data;
 };
